create form student for page infos
