+++ src/routes/index.js
@@ -7,38 +7,38 @@ import Register from '../containers/register';
 import Login from '../containers/login';
 import PageShell from '../containers/components/PageShell/pageShell';
 import Home from '../containers/home';
-import PersonalInfos from '../containers/personalInfos';
+import PersonalInfos from '../containers/PersonalInfos/personalInfos';
 import SearchMonitoring from '../containers/searchMonitoring';
 import Results from '../containers/results';
 
 
-export default class App extends React.Component {    
+export default class App extends React.Component {
     render() {
         return (
             <Router>
                 <PublicRoute
                     exact path="/cadastro"
-                    component={ Register }
+                    component={Register}
                 />
                 <PublicRoute
                     exact path="/entrar"
-                    component={ Login }
+                    component={Login}
                 />
                 <PrivateRoute
                     exact path="/"
-                    component={ PageShell(Home) }
+                    component={PageShell(Home)}
                 />
                 <PrivateRoute
                     exact path="/home"
-                    component={ PageShell(Home) }
+                    component={PageShell(Home)}
                 />
                 <PrivateRoute
                     exact path="/personal-infos"
-                    component={ PageShell(PersonalInfos) }
+                    component={PageShell(PersonalInfos)}
                 />
                 <PrivateRoute
                     exact path="/search-monitoring"
-                    component={ PageShell(SearchMonitoring) }
+                    component={PageShell(SearchMonitoring)}
                 />
                 <PrivateRoute
                     exact path="/results"
